Backend/requirements.txt
Django==2.2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0